refactor: move UserInfo to separate file
